800 rating
It's a bit difficult question, so it needs some explanation on how to solve it:

Ilia calls every n minutes. If n = 1, Ilia calls at the 1st, 2nd, 3rd... minutes until the day ends at the zth minute.

Artists come every m minutes. If m = 2, artists arrive at the 2nd, 4th, 6th... minutes until the day ends at the zth minute.

If the minute when Ilia calls is the same as the minute when artists arrive, then that artist has to be "killed." In other words, we need to find how many times Ilia's calls overlap with the times artists arrive.

Since both Ilia's calls and the artists' arrivals occur at regular intervals (multiples of their respective n and m), we can find the overlaps using their least common multiple (LCM).


Steps to Solve the Problem:

Calculate the LCM of n and m: This will give us the minute when both Ilia's call and an artist's arrival coincide.

Count the occurrences of the LCM within the range from 1 to z. This tells us how many times the overlap happens during the day.

The final result will be the number of artists that need to be "killed" to ensure there are no artists in the room when Ilia calls.